requirements.txt
google-cloud-aiplatform
google-cloud-storage
vertexai
github
pygithub
google-cloud-secret-manager
